backend/requirements.txt
psycopg2-binary
gunicorn
django
djangorestframework
pytest
pytest-mock
uvicorn
pytest-django
django-extensions
pytest-asyncio
adrf
faker
pydevd-pycharm